enhancing the backend server with proper file structures and making updates to the overall code base
