+++ .env.example
@@ -1,1 +1,2 @@
 GROQ_API_KEY = 
+REDIS_URL =
